fix: Change notification alarm manager
Change AlarmManager.setRepeating to AlarmManager.set method as alarm triggers everyday at the set time even without the repeating parameter. AlarmManager.setExactAndAllowWhileIdle is used for SDK >= 23 because alarm was not triggered when the device is in Doze mode. For SDK versions lower than 23 but higher than 19, setExact method is used as Doze mode was not introduced in those versions.