refactor: Create Utils.java and move updateDaysInFridge method
Previous version of code only updated days to expiration at the set time only right before the notification shows. Thus, when users open the app before the set time, days to expiration were not updated. In order not to repeat the code, Utils.java is created with updateDaysInFridge method and move the code for updating days to expiration from notification intent service to this java file for future reuse.